requirements.txt
Fabric==1.10.0
Flask==0.10.1
Flask-Actions==0.6.6
Flask-Assets==0.10
Flask-Cache==0.13.1
Flask-Classy==0.6.10
Flask-DebugToolbar==0.9.0
Flask-Mail==0.9.0
Flask-Script==2.0.5
Flask-WTF==0.10.2
Flask-WebTest==0.0.7
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
PyYAML==3.11
Pycco==0.3.0
Pygments==1.6
WTForms==2.0.1
WebOb==1.4
WebTest==2.0.16
Werkzeug==0.9.6
argh==0.26.1
beautifulsoup4==4.3.2
blinker==1.3
bpython==0.13.1
coverage==3.7.1
cssmin==0.2.0
ecdsa==0.11
fabtools==0.19.0
ghp-import==0.4.1
gnureadline==6.3.3
gunicorn==19.1.1
ipython==2.2.0
itsdangerous==0.24
jsmin==2.0.11
mock==1.0.1
nose==1.3.4
paramiko==1.15.1
pathtools==0.1.2
pycrypto==2.6.1
pystache==0.5.4
readline==6.2.4.1
redis==2.10.3
requests==2.4.3
rethinkdb==1.15.0-1
six==1.8.0
smartypants==1.8.6
subprocess32==3.2.6
termcolor==1.1.0
waitress==0.8.9
watchdog==0.8.2
webassets==0.10.1
wsgiref==0.1.2
